EncodingUtils: removed pointless cast in UTF8Utils::decode().

+++ duct/impl/EncodingUtils.inl
@@ -80,9 +80,9 @@ RandomAccessIt UTF8Utils::decode(
 		0x00000000, 0x00003080, 0x000E2080,
 		0x03C82080, 0xFA082080, 0x82082080
 	};
-	unsigned const trailing_units = static_cast<std::uint8_t>(
-		s_utf8_trailing[static_cast<char8_strict>(*pos)]
-	);
+	unsigned const trailing_units = s_utf8_trailing[
+		static_cast<char8_strict>(*pos)
+	];
 	if (end > pos + trailing_units) {
 		output = 0;
 		switch (trailing_units) {
